Move DebugPass to separate file
